restructured the directories; removed some more redundant navbars and footers
