+++ loyola_website/components/home/Announcement_marquee.js
