+++ loyola_website/pages/index.js
@@ -1,11 +1,11 @@
-import Announcement_marquee from "components/shashank-announcement-marquee/Announcement_marquee";
+import Announcement_marquee from "@/components/home/Announcement_marquee";
 import Navbar from "../components/navbar/Navbar";
 import Footer from "../components/footer/Footer";
-import Courses from "../components/Courses";
-import WelcomeImage from "../components/WelcomeImage";
-import WelcomeText from "../components/WelcomeText";
-import SupportAssistance from "../components/SupportAssistance";
-import FeaturedRecentEvents from '../components/FeaturedRecentEvents';
+import Courses from "../components/home/Courses";
+import WelcomeImage from "../components/home/WelcomeImage";
+import WelcomeText from "../components/home/WelcomeText";
+import SupportAssistance from "../components/home/SupportAssistance";
+import FeaturedRecentEvents from '../components/home/FeaturedRecentEvents';
 
 import { Inter } from "next/font/google";
 const inter = Inter({ subsets: ['latin'] })
