--- loyola_website/pages/test.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import WelcomeText from '@/components/WelcomeText.js'
-const test = () => {
-  return (
-    <div><WelcomeText/></div>
-  )
-}
-
-export default test
