--- loyola_website/pages/welcomeTesting.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import WelcomeText from '@/components/WelcomeText'
-
-const welcomeTesting = () => {
-  return (
-    <WelcomeText/>
-  )
-}
-
-export default welcomeTesting
